Switch to igor2 for .ibw and unpeg NumPy
Closes #682

We had pegged `numpy==1.23.4` because the unmaintained [igor](https://pypi.org/project/igor/) used the now deprecated `np.complex`. By switching to the continuation project [igor2](https://pypi.org/project/igor2/) which still sees active maintenance we are able to remove the fixed numpy version.

I've also taken the liberty to remove `flake8` packages from the list of `dev` optional-dependencies as we no longer have `pre-commit` hooks in place and the task is done by `ruff` and added `ipython` as its a commonly used alternative to plain Python interface (I prefer it as it has additional/commands but always forget to install it in new VirtualEnvs I setup for testing).