Adds create-config entry point
Closes #757

Adds a `topostats create-config` entry point and removes reference to `run_topostats`.

As we move towards a new release having a modular interface is desirable and that includes generation of sample
configuration files. This commit brings the "swiss army knife" approach of `topostats <subcommand>`to parity with the
older `run_topostats` by adding `topostats create-config` to generate configuration files for users to edit.

Documentation has been updated to reflect these changes, including the `README.md`. Because we have versioned
documentation users will still be able to refer to the older documentation.

For now the `run_topostats` method remains but raises a deprecation warning if the `--create-config-file` option is used.

**NB** This work was undertaken as a consequence of re-writing the `.github/ISSUE_TEMPLATES/` so that the instructions
could use this new entry point (see separate Pull Request).