Sets the type of all arguments that are files/paths to Path
I noticed that the type of arguments could be set to Path and so have updated our entry_point argparse to use this.

This also required a small change to convert `Path() > str` when creating output files and tweaks to the tests to ensure
targets were of type `Path()`. Whilst updating these I noticed that some of the targets had spaces at the start, this
was caused by flags and arguments being combined in the list that are passed (i.e. `"-c default_config.yaml"`) has been
corrected so that they are separate items in the list (i.e. `"-c", "default_config.yaml"`) and the spaces at the start
of the targets are no longer required.