Ensures target directories exist
In #806 and refactoring (see 07db71820) I found an unintended side-effect was that some of the target
directories (specifically `filter_out_path` and `grain_out_path`) were no longer present and the
tests (`tests/test_processing.py`) failed.

These are meant to be explicitly created by `processing.get_out_paths()` but for some reason were now no longer.

For filters the solution was to ensure the `filter_out_path` is created _and_ use either the `core_out_path` or
`filter_out_path` depending on which image is being saved.

For grains the `direction` is appended to the `grain_out_path` once (rather than three times) and the `.mkdir(parents =
True, exists_ok=True)` pathlib method is used to ensure the directory exists prior to attempting to plot.

In simplifying the logic I found that some of the processing tests failed, for some reason images that were meant to be
part of the `all` set were being plotted by mistake and the tests were looking for these images in the `caplog`. These
have been updated.