Just return min/max feret distances and coordinates
After discussion with @SylivaWhittle and @llwiggins it was decided that because of the issue of minimum feret
coordinates sometimes falling outside of the convex hull that we should not use it as a consistent line through the
grain/image.

Instead we will use the maximum feret and a line perpendicular to the mid-point of the maximum feret.

As such all attempts to work out if minimum feret coordinates/lines were inside or on the edge of a polygon have now
been removed and the function returns a dictionary of...

+ `min_feret`
+ `min_feret_coords`
+ `max_feret`
+ `max_feret_coords`

The dictionary should be conducive to being saved in the HDF5 output (perhaps a separate issue to do so).