refactor(nodestats): grain > mask; smoothed_grain > smoothed_mask
Converting use of `grain` to `mask` in `nodestats` to align with nomenclature in `dnatracing`

Its more consistent and descriptive of what the arrays are in terms of NumPy parlance (they are all binary masks) and
avoids confusion where at times in the variables labelled `grain` historically contained height values and more recently
it contains masked data that is binary.

Improves type hints for numpy arrays.