bug(grainstats): Improved bounding of aspect ratio to be 0 to 1
Rather than calculating `aspect_ratio` and as the ratio of the length on the x-axis divided by the length of the y-axis
a simpler method is possible.

Because we calculated the `smallest_bounding_width` and `smallest_bounding_length` then former divided by the later will
always be in the range of 0 to 1 as desired.

Updates regression test targets in light of this change.