tests(unet_masking): Correct ids and add expected fail
- Corrects the `id` of the parameterised tests for `test_make_bounding_box_square()` to reflect the `row`/`col`
  and whether they `increase`/`decrease`.
- Adds a rectangular image with a large rectangular `bbox` that can not be squared and marks the test as a fail.

I recall, but can't find, that we added functionality to handle non-square scans (I think @SylviaWhittle wrote the code
to handle it) and so its possible that instances may arise where the `bbox` can't be squared.

Ideally we should add a `try: ... except:` to handle such cases but for now adding a test which is marked as expected to
fail should serve as an _aide memoire_ that sometimes things won't work.