refactor(disordered_tracing): Work with skan-0.12.2
Closes #986

With the release of `skan-0.12.2` we can remove the code which converts `-` in the variable names of Pandas DataFrames
to instead use `_` which allows the use of "dot-notation" when referring to the variables of data frames.

We therefore unpin the `skan` version in `pyproject.toml`, explicitly state the `separator="_"` argument to
`skan.summarize()` and adapt the hard-coded variable names appropriately. As noted in the [release
notes](https://github.com/jni/skan/releases/tag/v0.12.2) the default will be `_` from `v0.13.0` and the flag allows
backwards compatibility but since we have gone with the use of `_` in variable names it will do no harm leaving these in
place (although they could be removed when `skan-v0.13.0` is released).