feature: Sub-command to run filtering
Closes #741

Adds the "swiss-army knife" component to run just filtering on files.

This involved modifying how the `.topostats` files are loaded and extracted because of the nesting (see #1068).

Tests currently fail because the `tests/resources/test_image/minicircle_small.topostats` is version `0.1` and doesn't
therefore work with the refactored structure (surprised #1068 passed all tests actualy!).

A separate commit will be made for updating this test file and the associated tweaking of tests.