feature: Adds sub-command to run grains step
Closes #742

- Adds `processing.process_grains()`
- Adds `run_modules.grains()`

Together these allow the `topostats grains` entry point to run which loads `*.topostats` files (`v0.2`), extracts the
flattened image and re-runs grains.

If any processing artefacts from previous runs are present in the `.topostats` files they are removed. Output is written
to the specified directory so that comparisions will be possible.

Added in tests and checked that some things in the previous `topstats filters` step work too. Some code is in place for
subsequent modules that will be added in turn and they may need refining.