docs: how to add functions and modules
Closes #989

As I go through writing additional entry points for #517 I thought it worth documenting how to incorporate functionality
into the workflow in terms of adding it to the end-to-end pipeline and how to create an additional entry point so that
the specific functionality can be run on its own.

This is based on having done `filters` (#741) over a few weeks and as I go forward with subsequent (e.g. #742) which I've
started I may well tweak/revise these notes in case I've missed anything or made mistakes.