feature(grains|io): Method to convert GrainCrop to dictionary
Adds a method to convert a `grains.GrainCrop` object to a dictionary (`grains.GrainCrop.grain_crop_to_dict()`) which is
indexed by the attribute names with the attributes as the values. Includes a simple test using the `dummy_graincrop`
fixture.

As this will be needed to convert the nested dictionaries of `ImageGrainCrops` (which hold `GrainCrop` objects within
`GrainCropsDirection`) objects we include a check to see if the `item` (value) of a dictionary is of type `GrainCrop`
and if so recursively call `dict_to_hdf5` on the result of `item.grain_crop_to_dict()`. Again a simple test to ensure
this functions as expected is included.