feature(grains|io): GrainCropsDirection to dictionary
**NB** This is a stacked pull request and should be merged _after_ #1104

Adds a method, based on the `__dict__` "magic dunder" method to convert the `GrainCropsDirection` to a dictionary. Tests
are included not just for the method itself but because this is required in `io.dict_to_hdf5` there are also tests that
it works with that (these are contingent on changes introduced in #1104 because `GrainCrop` are nested within the
`GrainCropsDirection` object).